Move copied-and-pasted code into a single header file
The code that blocks on `/dev/random` on Linux to ensure that
`/dev/urandom` is properly seeded was duplicated in
src/libsodium/randombytes/sysrandom/randombytes_sysrandom.c and in
src/libsodium/randombytes/salsa20/randombytes_salsa20_random.c.  Move it
into a single header file.

Also, update .gitignore to ignore Emacs temporary files.
